refactor to handle compat with numpy 1/2
